Add additional warnings: write-strings and conversion

See merge request libtiff/libtiff!802

+++ contrib/dbs/tiff-bi.c
@@ -67,7 +67,7 @@ int main(int argc, char **argv)
         scan_line[i] = 255;
 
     for (i = 0; i < HEIGHT / 2; i++)
-        TIFFWriteScanline(tif, scan_line, i, 0);
+        TIFFWriteScanline(tif, scan_line, (uint32_t)i, 0);
 
     for (i = 0; i < (WIDTH / 8) / 2; i++)
         scan_line[i] = 255;
@@ -76,7 +76,7 @@ int main(int argc, char **argv)
         scan_line[i] = 0;
 
     for (i = HEIGHT / 2; i < HEIGHT; i++)
-        TIFFWriteScanline(tif, scan_line, i, 0);
+        TIFFWriteScanline(tif, scan_line, (uint32_t)i, 0);
 
     free(scan_line);
     TIFFClose(tif);
