+++ contrib/dbs/tiff-grayscale.c
@@ -74,14 +74,14 @@ int main(int argc, char **argv)
     }
 
     cmsize = nchunks * nchunks;
-    gray = (uint16_t *)malloc(cmsize * sizeof(uint16_t));
+    gray = (uint16_t *)malloc((size_t)cmsize * sizeof(uint16_t));
 
     gray[0] = 3000;
     for (i = 1; i < cmsize; i++)
         gray[i] = (uint16_t)(-log10((double)i / (cmsize - 1)) * 1000);
 
     refblackwhite[0] = 0.0;
-    refblackwhite[1] = (float)((1L << bits_per_pixel) - 1);
+    refblackwhite[1] = (float)((1UL << bits_per_pixel) - 1);
 
     if ((tif = TIFFOpen(argv[3], "w")) == NULL)
     {
@@ -102,7 +102,7 @@ int main(int argc, char **argv)
     TIFFSetField(tif, TIFFTAG_TRANSFERFUNCTION, gray);
     TIFFSetField(tif, TIFFTAG_RESOLUTIONUNIT, RESUNIT_NONE);
 
-    scan_line = (unsigned char *)malloc(WIDTH / (8 / bits_per_pixel));
+    scan_line = (unsigned char *)malloc((size_t)(WIDTH / (8 / bits_per_pixel)));
 
     for (i = 0; i < HEIGHT; i++)
     {
@@ -131,7 +131,7 @@ int main(int argc, char **argv)
                     break;
             }
         }
-        TIFFWriteScanline(tif, scan_line, i, 0);
+        TIFFWriteScanline(tif, scan_line, (uint32_t)i, 0);
     }
 
     free(scan_line);
